+++ Sources/KeyboardKit/Actions/KeyboardAction+Actions.swift
@@ -23,13 +23,13 @@ public extension KeyboardAction {
     
     /**
      The action that by default should be triggered when the
-     keyboard action is double tapped.
+     action is double tapped.
      */
     var standardDoubleTapAction: GestureAction? { nil }
     
     /**
      The action that by default should be triggered when the
-     keyboard action is long pressed.
+     action is long pressed.
      */
     var standardLongPressAction: GestureAction? {
         switch self {
@@ -40,7 +40,7 @@ public extension KeyboardAction {
     
     /**
      The action that by default should be triggered when the
-     keyboard action is pressed down.
+     action is pressed.
      */
     var standardPressAction: GestureAction? {
         switch self {
@@ -51,13 +51,13 @@ public extension KeyboardAction {
     
     /**
      The action that by default should be triggered when the
-     keyboard action is released.
+     action is released.
      */
     var standardReleaseAction: GestureAction? { nil }
     
     /**
      The action that by default should be triggered when the
-     keyboard action pressed and hold.
+     action is pressed, until it's released.
      */
     var standardRepeatAction: GestureAction? {
         switch self {
@@ -68,27 +68,52 @@ public extension KeyboardAction {
     
     /**
      The action that by default should be triggered when the
-     keyboard action is tapped.
+     action is tapped.
      */
     var standardTapAction: GestureAction? {
-        if self.isPrimaryAction { return { $0?.textDocumentProxy.insertText("\n") }}
+        if let action = standardTextDocumentProxyAction { return action }
         
         switch self {
-        case .backspace: return { $0?.textDocumentProxy.deleteBackward() }
-        case .character(let char): return { $0?.textDocumentProxy.insertText(char) }
         case .dismissKeyboard: return { $0?.dismissKeyboard() }
-        case .emoji(let emoji): return { $0?.textDocumentProxy.insertText(emoji.char) }
-        case .moveCursorBackward: return { $0?.textDocumentProxy.adjustTextPosition(byCharacterOffset: -1) }
-        case .moveCursorForward: return { $0?.textDocumentProxy.adjustTextPosition(byCharacterOffset: 1) }
         case .nextLocale: return { $0?.keyboardContext.selectNextLocale() }
-        case .newLine: return { $0?.textDocumentProxy.insertText("\n") }
-        case .return: return { $0?.textDocumentProxy.insertText("\n") }
         case .shift(let currentState): return {
             switch currentState {
             case .lowercased, .neutral: $0?.keyboardContext.keyboardType = .alphabetic(.uppercased)
             case .uppercased, .capsLocked: $0?.keyboardContext.keyboardType = .alphabetic(.lowercased)
             }
         }
+        default: return nil
+        }
+    }
+    
+    /**
+     The standard text document proxy action, if any.
+     */
+    var standardTextDocumentProxyAction: GestureAction? {
+        if let action = standardTextDocumentProxyInputAction { return action }
+        
+        switch self {
+        case .moveCursorBackward: return { $0?.textDocumentProxy.adjustTextPosition(byCharacterOffset: -1) }
+        case .moveCursorForward: return { $0?.textDocumentProxy.adjustTextPosition(byCharacterOffset: 1) }
+        default: return nil
+        }
+    }
+    
+    /**
+     The standard text document proxy input action, if any.
+     
+     An input action either inserts or deletes text from the
+     text document proxy.
+     */
+    var standardTextDocumentProxyInputAction: GestureAction? {
+        if self.isPrimaryAction { return { $0?.textDocumentProxy.insertText("\n") }}
+        
+        switch self {
+        case .backspace: return { $0?.textDocumentProxy.deleteBackward() }
+        case .character(let char): return { $0?.textDocumentProxy.insertText(char) }
+        case .emoji(let emoji): return { $0?.textDocumentProxy.insertText(emoji.char) }
+        case .newLine: return { $0?.textDocumentProxy.insertText("\n") }
+        case .return: return { $0?.textDocumentProxy.insertText("\n") }
         case .space: return { $0?.textDocumentProxy.insertText(" ") }
         case .tab: return { $0?.textDocumentProxy.insertText("\t") }
         default: return nil