+++ Tests/KeyboardKitTests/Actions/KeyboardAction+ActionsTests.swift
@@ -34,23 +34,35 @@ class KeyboardAction_ActionsTests: QuickSpec {
         
         describe("standard double tap action") {
             
+            func result(for action: KeyboardAction) -> Any? {
+                action.standardDoubleTapAction
+            }
+            
             it("is not defined for any actions") {
                 expected = []
-                expected.forEach { expect($0.standardDoubleTapAction).toNot(beNil()) }
-                unexpected.forEach { expect($0.standardDoubleTapAction).to(beNil()) }
+                expected.forEach { expect(result(for: $0)).toNot(beNil()) }
+                unexpected.forEach { expect(result(for: $0)).to(beNil()) }
             }
         }
         
         describe("standard long press action") {
             
+            func result(for action: KeyboardAction) -> Any? {
+                action.standardLongPressAction
+            }
+            
             it("is defined for some actions") {
                 expected = [.backspace]
-                expected.forEach { expect($0.standardLongPressAction).toNot(beNil()) }
-                unexpected.forEach { expect($0.standardLongPressAction).to(beNil()) }
+                expected.forEach { expect(result(for: $0)).toNot(beNil()) }
+                unexpected.forEach { expect(result(for: $0)).to(beNil()) }
             }
         }
         
         describe("standard press action") {
+            
+            func result(for action: KeyboardAction) -> Any? {
+                action.standardPressAction
+            }
             
             it("is not defined for any actions") {
                 expected = [
@@ -64,21 +76,29 @@ class KeyboardAction_ActionsTests: QuickSpec {
                     .keyboardType(.images),
                     .keyboardType(.custom(""))
                 ]
-                expected.forEach { expect($0.standardPressAction).toNot(beNil()) }
-                unexpected.forEach { expect($0.standardPressAction).to(beNil()) }
+                expected.forEach { expect(result(for: $0)).toNot(beNil()) }
+                unexpected.forEach { expect(result(for: $0)).to(beNil()) }
             }
         }
         
         describe("standard release action") {
             
+            func result(for action: KeyboardAction) -> Any? {
+                action.standardReleaseAction
+            }
+            
             it("is not defined for any actions") {
                 expected = []
-                expected.forEach { expect($0.standardReleaseAction).toNot(beNil()) }
-                unexpected.forEach { expect($0.standardReleaseAction).to(beNil()) }
+                expected.forEach { expect(result(for: $0)).toNot(beNil()) }
+                unexpected.forEach { expect(result(for: $0)).to(beNil()) }
             }
         }
         
         describe("standard tap action") {
+            
+            func result(for action: KeyboardAction) -> Any? {
+                action.standardTapAction
+            }
             
             it("is defined for some actions") {
                 expected = [
@@ -101,17 +121,73 @@ class KeyboardAction_ActionsTests: QuickSpec {
                     .space,
                     .tab
                 ]
-                expected.forEach { expect($0.standardTapAction).toNot(beNil()) }
-                unexpected.forEach { expect($0.standardTapAction).to(beNil()) }
+                expected.forEach { expect(result(for: $0)).toNot(beNil()) }
+                unexpected.forEach { expect(result(for: $0)).to(beNil()) }
+            }
+        }
+        
+        describe("standard text document proxy action") {
+            
+            func result(for action: KeyboardAction) -> Any? {
+                action.standardTextDocumentProxyAction
+            }
+            
+            it("is defined for some actions") {
+                expected = [
+                    .backspace,
+                    .character(""),
+                    .emoji(Emoji("")),
+                    .moveCursorBackward,
+                    .moveCursorForward,
+                    .newLine,
+                    .primary(.done),
+                    .primary(.go),
+                    .primary(.ok),
+                    .primary(.search),
+                    .return,
+                    .space,
+                    .tab
+                ]
+                expected.forEach { expect(result(for: $0)).toNot(beNil()) }
+                unexpected.forEach { expect(result(for: $0)).to(beNil()) }
+            }
+        }
+        
+        describe("standard text document proxy input action") {
+            
+            func result(for action: KeyboardAction) -> Any? {
+                action.standardTextDocumentProxyInputAction
+            }
+            
+            it("is defined for some actions") {
+                expected = [
+                    .backspace,
+                    .character(""),
+                    .emoji(Emoji("")),
+                    .newLine,
+                    .primary(.done),
+                    .primary(.go),
+                    .primary(.ok),
+                    .primary(.search),
+                    .return,
+                    .space,
+                    .tab
+                ]
+                expected.forEach { expect(result(for: $0)).toNot(beNil()) }
+                unexpected.forEach { expect(result(for: $0)).to(beNil()) }
             }
         }
         
         describe("standard repeat action") {
             
+            func result(for action: KeyboardAction) -> Any? {
+                action.standardRepeatAction
+            }
+            
             it("is defined for some actions") {
                 expected = [.backspace]
-                expected.forEach { expect($0.standardRepeatAction).toNot(beNil()) }
-                unexpected.forEach { expect($0.standardRepeatAction).to(beNil()) }
+                expected.forEach { expect(result(for: $0)).toNot(beNil()) }
+                unexpected.forEach { expect(result(for: $0)).to(beNil()) }
             }
         }
     }